# PrASeguRed/src/main/resources/static/openapi.yaml
openapi: 3.0.0
info:
  title: API Gestión Usuarios y Reportes
  description: |
    API para la gestión de usuarios, autenticación, reportes y notificaciones de la plataforma PRA-Segured.
    Proporciona endpoints para registro, verificación, manejo de reportes ciudadanos y sistema de comentarios.
  version: 1.6.3
  contact:
    name: Equipo PRA-Segured
    email: soporte@prasegured.co
servers:
  - url: /
    description: Servidor de desarrollo local
  - url: https://localhost:8080
    description: Servidor de pruebas con HTTPS
components:
  schemas:
...
    Categoria:
      type: object
      required:
        - nombre
      properties:
        nombre:
          type: string
          example: "Infraestructura vial"
        descripcion:
          type: string
          example: "Problemas relacionados con calles, avenidas y señalización"
    ESTADOREPORTE:
      type: string
      enum:
        - Pendiente
        - Completado
        - Denegado
        - Eliminado
        - Espera
      description: Estado actual del reporte en el sistema
...
    ReporteDTO:
      type: object
      properties:
        id:
          type: string
          description: ID único del reporte
          example: "60d85c3e7b4a9d001ff2a456"
        idUsuario:
          type: string
          description: ID del usuario que creó el reporte
          example: "60d85c3e7b4a9d001ff2a123"
        estado:
          $ref: '#/components/schemas/ESTADOREPORTE'
          example: "Pendiente"
        titulo:
          type: string
          description: Título del reporte
          example: "Hueco peligroso en la vía principal"
        fechaPublicacion:
          type: string
          format: date-time
          description: Fecha en que se publicó el reporte
          example: "2025-05-15T15:29:53.000Z"
        fechaActualizacion:
          type: string
          format: date-time
          description: Fecha de la última actualización del reporte
          example: "2025-05-16T08:45:12.000Z"
        descripcion:
          type: string
          description: Descripción detallada del reporte
          example: "Hay un hueco de aproximadamente 1 metro de diámetro y 20 cm de profundidad en la Av. Bolívar frente al centro comercial. Ya ha causado accidentes."
        likes:
          type: integer
          description: Número de "me gusta" que tiene el reporte
          example: 15
        categoria:
          type: array
          items:
            $ref: '#/components/schemas/Categoria'
          description: Categorías a las que pertenece el reporte
        location:
          $ref: '#/components/schemas/Location'
          description: Ubicación del incidente reportado
        imagenes:
          type: array
          items:
            $ref: '#/components/schemas/Imagen'
          description: Imágenes asociadas al reporte
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtenido durante el proceso de autenticación
paths:
...
  /reportes:
    get:
      summary: Obtener todos los reportes
      description: Recupera una lista de todos los reportes disponibles
      operationId: obtenerTodosReportes
      security:
        - bearerAuth: []
      parameters:
        - name: estado
          in: query
          description: Filtrar por estado de reporte
          required: false
          schema:
            type: string
            enum: [Pendiente, Completado, Denegado, Espera]
        - name: categoria
          in: query
          description: Filtrar por categoría
          required: false
          schema:
            type: string
        - name: page
          in: query
          description: Número de página para paginación
          required: false
          schema:
            type: integer
            default: 0
        - name: size
          in: query
          description: Tamaño de página para paginación
          required: false
          schema:
            type: integer
            default: 10
      responses:
        "200":
          description: Lista de reportes obtenida exitosamente
          content:
            application/json:
              schema:
                type: object
                properties:
                  content:
                    type: array
                    items:
                      $ref: '#/components/schemas/ReporteDTO'
                  totalPages:
                    type: integer
                    example: 5
                  totalElements:
                    type: integer
                    example: 42
                  last:
                    type: boolean
                    example: false
                  size:
                    type: integer
                    example: 10
        "401":
          description: No autenticado o token inválido
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Token inválido o expirado."

    post:
      summary: Creación reporte
      description: Creación de un nuevo reporte en el sistema
      operationId: reporteCreacion
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReporteRequest'
            examples:
              nuevoReporte:
                value:
                  titulo: "Hueco peligroso en la vía principal"
                  descripcion: "Hay un hueco de aproximadamente 1 metro de diámetro y 20 cm de profundidad en la Av. Bolívar frente al centro comercial. Ya ha causado accidentes a motociclistas y biciusuarios."
                  categoria: [
                    {
                      "nombre": "Infraestructura vial",
                      "descripcion": "Problemas relacionados con calles y carreteras"
                    },
                    {
                      "nombre": "Seguridad",
                      "descripcion": "Situaciones que ponen en riesgo a los ciudadanos"
                    }
                  ]
                  locations: {
                    "latitude": 4.550923,
                    "longitude": -75.656080,
                    "name": "Av. Bolívar frente al Portal del Quindío"
                  }
                  imagenes: [
                    {
                      "url": "data:image/jpeg;base64,/9j/4AAQSkZJRgABAQEAYABgAAD/4QBoRXhpZgAATU0AKgAAAAgABAEaAAUAAAABAAAAPgEbAAUAAAABAAAARgEoAAMAAAABAAIAAAExAAIAAAASAAAATgAAAAAAAABgAAAAAQAAAGAAAAABUGFpbnQuTkVUIHYzLjUuMTAA/9sAQwAEAgMDAwIEAwMDBAQEBAUJBgUFBQULCAgGCQ0LDQ0NCwwMDhAUEQ4PEw8MDBIYEhMVFhcXFw4RGRsZFhoUFhcW/9sAQwEEBAQFBQUKBgYKFg8MDxYWFhYWFhYWFhYWFhYWFhYWFhYWFhYWFhYWFhYWFhYWFhYWFhYWFhYWFhYWFhYWFhYW/8AAEQgAZABkAwEiAAIRAQMRAf/EAB8AAAEFAQEBAQEBAAAAAAAAAAABAgMEBQYHCAkKC//EALUQAAIBAwMCBAMFBQQEAAABfQECAwAEEQUSITFBBhNRYQcicQgUKjKBCRVCsRIjM2JyggkKFhcYGRolJicoKSo0NTY3ODk6Q0RFRkdISUpTVFVWV1hZWmNkZWZnaGlqc3R1dnd4eXqDhIWGh4iJipKTlJWWl5iZmqKjpKWmp6ipqrKztLW2t7i5usLDxMXGx8jJytLT1NXW19jZ2uHi4+Tl5ufo6erx8vP09fb3+Pn6/8QAHwEAAwEBAQEBAQEBAQAAAAAAAAECAwQFBgcICQoL/8QAtREAAgECBAQDBAcFBAQAAQJ3AAECAxEEBSExBhJBUQdhcRMiMoEIFEKRobHBCSMzUvAVYnLRChYkNOEl8RcYGRomJygpKjU2Nzg5OkNERUZHSElKU1RVVldYWVpjZGVmZ2hpanN0dXZ3eHl6goOEhYaHiImKkpOUlZaXmJmaoqOkpaanqKmqsrO0tba3uLm6wsPExcbHyMnK0tPU1dbX2Nna4uPk5ebn6Onq8vP09fb3+Pn6/9oADAMBAAIRAxEAPwD7LooooAKKKKACiiigAooooAKKKOtABRTWdUGWIUe9Y2s+KtL0aVYbiZmmf7sUYLE/l0/GplOMF7zLhTnUdoq5uVXv...",
                      "descripcion": "Vista del hueco desde la acera"
                    },
                    {
                      "url": "data:image/jpeg;base64,/9j/4AAQSkZJRgABAQEAYABgAAD/4QBoRXhpZgAATU0AKgAAAAgABAEaAAUAAAABAAAAPgEbAAUAAAABAAAARgEoAAMAAAABAAIAAAExAAIAAAASAAAATgAAAAAAAABgAAAAAQAAAGAAAAABUGFpbnQuTkVUIHYzLjUuMTAA/9sAQwAEAgMDAwIEAwMDBAQEBAUJBgUFBQULCAgGCQ0LDQ0NCwwMDhAUEQ4PEw8MDBIYEhMVFhcXFw4RGRsZFhoUFhcW/9sAQwEEBAQFBQUKBgYKFg8MDxYWFhYWFhYWFhYWFhYWFhYWFhYWFhYWFhYWFhYWFhYWFhYWFhYWFhYWFhYWFhYWFhYW/8AAEQgAZABkAwEiAAIRAQMRAf/EAB8AAAEFAQEBAQEBAAAAAAAAAAABAgMEBQYHCAkKC//EALUQAAIBAwMCBAMFBQQEAAABfQECAwAEEQUSITFBBhNRYQcicQgUKjKBCRVCsRIjM2JyggkKFhcYGRolJicoKSo0NTY3ODk6Q0RFRkdISUpTVFVWV1hZWmNkZWZnaGlqc3R1dnd4eXqDhIWGh4iJipKTlJWWl5iZmqKjpKWmp6ipqrKztLW2t7i5usLDxMXGx8jJytLT1NXW19jZ2uHi4+Tl5ufo6erx8vP09fb3+Pn6/8QAHwEAAwEBAQEBAQEBAQAAAAAAAAECAwQFBgcICQoL/8QAtREAAgECBAQDBAcFBAQAAQJ3AAECAxEEBSExBhJBUQdhcRMiMoEIFEKRobHBCSMzUvAVYnLRChYkNOEl8RcYGRomJygpKjU2Nzg5OkNERUZHSElKU1RVVldYWVpjZGVmZ2hpanN0dXZ3eHl6goOEhYaHiImKkpOUlZaXmJmaoqOkpaanqKmqsrO0tba3uLm6wsPExcbHyMnK0tPU1dbX2Nna4uPk5ebn6Onq8vP09fb3+Pn6/9oADAMBAAIRAxEAPwD7LooooAKKKKACiiigAooooAKKKOtABRTWdUGWIUe9Y2s+KtL0aVYbiZmmf7sUYLE/l0/GplOMF7zLhTnUdoq5uVXv...",
                      "descripcion": "Medición del diámetro del hueco"
                    }
                  ]
      responses:
        "201":
          description: Reporte creado exitosamente.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Reporte creado exitosamente."
                  reporte:
                    $ref: '#/components/schemas/ReporteDTO'
        "401":
          description: No autenticado o token inválido
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Token inválido o expirado."
        "400":
          description: Datos del reporte inválidos o encabezado requerido faltante
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Solicitud inválida: datos insuficientes o incorrectos."
                  details:
                    type: array
                    items:
                      type: string
                    example:
                      - "El título no puede estar vacío"
                      - "La descripción debe tener al menos 20 caracteres"
  /reportes/{id}:
    get:
      summary: Obtener un reporte específico
      description: Recupera un reporte por su ID
      operationId: obtenerReportePorId
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            description: ID del reporte a obtener
            example: "60d85c3e7b4a9d001ff2a456"
      responses:
        "200":
          description: Reporte recuperado exitosamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReporteDTO'
        "401":
          description: No autenticado o token inválido
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Token inválido o expirado."
        "404":
          description: Reporte no encontrado
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Reporte no encontrado con ID: 60d85c3e7b4a9d001ff2a456"

    put:
      summary: Actualización reporte
      description: Modificación de un reporte existente
      operationId: reporteActualizacion
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            description: ID del reporte a actualizar
            example: "60d85c3e7b4a9d001ff2a456"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReporteRequest'
            examples:
              actualizarReporte:
                value:
                  titulo: "URGENTE: Hueco peligroso en la vía principal - Accidente reciente"
                  descripcion: "Hay un hueco de aproximadamente 1 metro de diámetro y 20 cm de profundidad en la Av. Bolívar frente al centro comercial. Actualización: Hoy hubo un accidente donde una motocicleta cayó en el hueco. Se necesita intervención urgente."
                  categoria: [
                    {
                      "nombre": "Infraestructura vial",
                      "descripcion": "Problemas relacionados con calles y carreteras"
                    },
                    {
                      "nombre": "Seguridad",
                      "descripcion": "Situaciones que ponen en riesgo a los ciudadanos"
                    },
                    {
                      "nombre": "Emergencia",
                      "descripcion": "Situación que requiere atención inmediata"
                    }
                  ]
                  locations: {
                    "latitude": 4.550923,
                    "longitude": -75.656080,
                    "name": "Av. Bolívar frente al Portal del Quindío"
                  }
                  imagenes: [
                    {
                      "url": "data:image/jpeg;base64,/9j/4AAQSkZJRgABAQEAYABgAAD/4QBoRXhpZgAATU0AKgAAAAgABAEaAAUAAAABAAAAPgEbAAUAAAABAAAARgEoAAMAAAABAAIAAAExAAIAAAASAAAATgAAAAAAAABgAAAAAQAAAGAAAAABUGFpbnQuTkVUIHYzLjUuMTAA/9sAQwAEAgMDAwIEAwMDBAQEBAUJBgUFBQULCAgGCQ0LDQ0NCwwMDhAUEQ4PEw8MDBIYEhMVFhcXFw4RGRsZFhoUFhcW/9sAQwEEBAQFBQUKBgYKFg8MDxYWFhYWFhYWFhYWFhYWFhYWFhYWFhYWFhYWFhYWFhYWFhYWFhYWFhYWFhYWFhYWFhYW/8AAEQgAZABkAwEiAAIRAQMRAf/EAB8AAAEFAQEBAQEBAAAAAAAAAAABAgMEBQYHCAkKC//EALUQAAIBAwMCBAMFBQQEAAABfQECAwAEEQUSITFBBhNRYQcicQgUKjKBCRVCsRIjM2JyggkKFhcYGRolJicoKSo0NTY3ODk6Q0RFRkdISUpTVFVWV1hZWmNkZWZnaGlqc3R1dnd4eXqDhIWGh4iJipKTlJWWl5iZmqKjpKWmp6ipqrKztLW2t7i5usLDxMXGx8jJytLT1NXW19jZ2uHi4+Tl5ufo6erx8vP09fb3+Pn6/8QAHwEAAwEBAQEBAQEBAQAAAAAAAAECAwQFBgcICQoL/8QAtREAAgECBAQDBAcFBAQAAQJ3AAECAxEEBSExBhJBUQdhcRMiMoEIFEKRobHBCSMzUvAVYnLRChYkNOEl8RcYGRomJygpKjU2Nzg5OkNERUZHSElKU1RVVldYWVpjZGVmZ2hpanN0dXZ3eHl6goOEhYaHiImKkpOUlZaXmJmaoqOkpaanqKmqsrO0tba3uLm6wsPExcbHyMnK0tPU1dbX2Nna4uPk5ebn6Onq8vP09fb3+Pn6/9oADAMBAAIRAxEAPwD7LooooAKKKKACiiigAooooAKKKOtABRTWdUGWIUe9Y2s+KtL0aVYbiZmmf7sUYLE/l0/GplOMF7zLhTnUdoq5uVXv...",
                      "descripcion": "Vista del hueco desde la acera"
                    },
                    {
                      "url": "data:image/jpeg;base64,/9j/4AAQSkZJRgABAQEAYABgAAD/4QBoRXhpZgAATU0AKgAAAAgABAEaAAUAAAABAAAAPgEbAAUAAAABAAAARgEoAAMAAAABAAIAAAExAAIAAAASAAAATgAAAAAAAABgAAAAAQAAAGAAAAABUGFpbnQuTkVUIHYzLjUuMTAA/9sAQwAEAgMDAwIEAwMDBAQEBAUJBgUFBQULCAgGCQ0LDQ0NCwwMDhAUEQ4PEw8MDBIYEhMVFhcXFw4RGRsZFhoUFhcW/9sAQwEEBAQFBQUKBgYKFg8MDxYWFhYWFhYWFhYWFhYWFhYWFhYWFhYWFhYWFhYWFhYWFhYWFhYWFhYWFhYWFhYWFhYW/8AAEQgAZABkAwEiAAIRAQMRAf/EAB8AAAEFAQEBAQEBAAAAAAAAAAABAgMEBQYHCAkKC//EALUQAAIBAwMCBAMFBQQEAAABfQECAwAEEQUSITFBBhNRYQcicQgUKjKBCRVCsRIjM2JyggkKFhcYGRolJicoKSo0NTY3ODk6Q0RFRkdISUpTVFVWV1hZWmNkZWZnaGlqc3R1dnd4eXqDhIWGh4iJipKTlJWWl5iZmqKjpKWmp6ipqrKztLW2t7i5usLDxMXGx8jJytLT1NXW19jZ2uHi4+Tl5ufo6erx8vP09fb3+Pn6/8QAHwEAAwEBAQEBAQEBAQAAAAAAAAECAwQFBgcICQoL/8QAtREAAgECBAQDBAcFBAQAAQJ3AAECAxEEBSExBhJBUQdhcRMiMoEIFEKRobHBCSMzUvAVYnLRChYkNOEl8RcYGRomJygpKjU2Nzg5OkNERUZHSElKU1RVVldYWVpjZGVmZ2hpanN0dXZ3eHl6goOEhYaHiImKkpOUlZaXmJmaoqOkpaanqKmqsrO0tba3uLm6wsPExcbHyMnK0tPU1dbX2Nna4uPk5ebn6Onq8vP09fb3+Pn6/9oADAMBAAIRAxEAPwD7LooooAKKKKACiiigAooooAKKKOtABRTWdUGWIUe9Y2s+KtL0aVYbiZmmf7sUYLE/l0/GplOMF7zLhTnUdoq5uVXv...",
                      "descripcion": "Medición del diámetro del hueco"
                    },
                    {
                      "url": "data:image/jpeg;base64,/9j/4AAQSkZJRgABAQEAYABgAAD/4QBoRXhpZgAATU0AKgAAAAgABAEaAAUAAAABAAAAPgEbAAUAAAABAAAARgEoAAMAAAABAAIAAAExAAIAAAASAAAATgAAAAAAAABgAAAAAQAAAGAAAAABUGFpbnQuTkVUIHYzLjUuMTAA/9sAQwAEAgMDAwIEAwMDBAQEBAUJBgUFBQULCAgGCQ0LDQ0NCwwMDhAUEQ4PEw8MDBIYEhMVFhcXFw4RGRsZFhoUFhcW/9sAQwEEBAQFBQUKBgYKFg8MDxYWFhYWFhYWFhYWFhYWFhYWFhYWFhYWFhYWFhYWFhYWFhYWFhYWFhYWFhYWFhYWFhYW/8AAEQgAZABkAwEiAAIRAQMRAf/EAB8AAAEFAQEBAQEBAAAAAAAAAAABAgMEBQYHCAkKC//EALUQAAIBAwMCBAMFBQQEAAABfQECAwAEEQUSITFBBhNRYQcicQgUKjKBCRVCsRIjM2JyggkKFhcYGRolJicoKSo0NTY3ODk6Q0RFRkdISUpTVFVWV1hZWmNkZWZnaGlqc3R1dnd4eXqDhIWGh4iJipKTlJWWl5iZmqKjpKWmp6ipqrKztLW2t7i5usLDxMXGx8jJytLT1NXW19jZ2uHi4+Tl5ufo6erx8vP09fb3+Pn6/8QAHwEAAwEBAQEBAQEBAQAAAAAAAAECAwQFBgcICQoL/8QAtREAAgECBAQDBAcFBAQAAQJ3AAECAxEEBSExBhJBUQdhcRMiMoEIFEKRobHBCSMzUvAVYnLRChYkNOEl8RcYGRomJygpKjU2Nzg5OkNERUZHSElKU1RVVldYWVpjZGVmZ2hpanN0dXZ3eHl6goOEhYaHiImKkpOUlZaXmJmaoqOkpaanqKmqsrO0tba3uLm6wsPExcbHyMnK0tPU1dbX2Nna4uPk5ebn6Onq8vP09fb3+Pn6/9oADAMBAAIRAxEAPwD7LooooAKKKKACiiigAooooAKKKOtABRTWdUGWIUe9Y2s+KtL0aVYbiZmmf7sUYLE/l0/GplOMF7zLhTnUdoq5uVXv...",
                      "descripcion": "Accidente de motocicleta causado por el hueco"
                    }
                  ]
      responses:
        "200":
          description: Reporte actualizado.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Reporte actualizado exitosamente."
                  reporte:
                    $ref: '#/components/schemas/ReporteDTO'
        "401":
          description: No autenticado o token inválido
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Token inválido o expirado."
        "404":
          description: Reporte no encontrado
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Reporte no encontrado con ID: 60d85c3e7b4a9d001ff2a456"

    delete:
      summary: Eliminación reporte
      description: Marca el reporte como eliminado (cambia su estado a Eliminado)
      operationId: reporteEliminacion
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            description: ID del reporte a eliminar
            example: "60d85c3e7b4a9d001ff2a456"
      responses:
        "200":
          description: Reporte eliminado correctamente.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Reporte eliminado correctamente."
                  id:
                    type: string
                    example: "60d85c3e7b4a9d001ff2a456"
        "401":
          description: No autenticado o token inválido
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Token inválido o expirado."
        "404":
          description: Reporte no encontrado
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Reporte no encontrado con ID: 60d85c3e7b4a9d001ff2a456"
  /reportes/{id}/completar:
    put:
      summary: Marcar reporte como completado
      description: Cambia el estado del reporte a Completado
      operationId: marcarReporteComoCompletado
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            description: ID del reporte a marcar como completado
            example: "60d85c3e7b4a9d001ff2a456"
      responses:
        "200":
          description: Reporte marcado como completado correctamente.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Reporte marcado como completado exitosamente."
                  reporte:
                    $ref: '#/components/schemas/ReporteDTO'
        "401":
          description: No autenticado o token inválido
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Token inválido o expirado."
        "404":
          description: Reporte no encontrado
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Reporte no encontrado con ID: 60d85c3e7b4a9d001ff2a456"
  /reportes/{id}/denegar:
    put:
      summary: Marcar reporte como denegado
      description: Cambia el estado del reporte a Denegado
      operationId: marcarReporteComoDenegado
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            description: ID del reporte a marcar como denegado
            example: "60d85c3e7b4a9d001ff2a456"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                motivo:
                  type: string
                  description: Motivo por el que se deniega el reporte
            examples:
              denegarReporte:
                value:
                  motivo: "El reporte no contiene información suficiente para ser procesado. Por favor, proporcione más detalles sobre la ubicación exacta."
      responses:
        "200":
          description: Reporte marcado como denegado correctamente.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Reporte marcado como denegado."
                  reporte:
                    $ref: '#/components/schemas/ReporteDTO'
        "401":
          description: No autenticado o token inválido
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Token inválido o expirado."
        "404":
          description: Reporte no encontrado
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Reporte no encontrado con ID: 60d85c3e7b4a9d001ff2a456"
...